requirements.txt
pandas
numpy
scikit-learn
matplotlib
pyyaml
torch
stable-baselines3
gymnasium
sentence-transformers
faiss-cpu
